data-streaming-with-kafka/requirements.txt
certifi==2021.5.30
chardet==4.0.0
cramjam==2.3.2
Faker==8.2.1
fsspec==2021.5.0
idna==2.10
kafka-python==2.0.2
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
text-unidecode==1.3
thrift==0.13.0
urllib3==1.26.5
